// src/__generated__/findChildCareProviders.ts
export interface findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_address {
  __typename: "Address";
  /**
   * Address line one containing house number
   */
  addressLine1: string;
  /**
   * Address line two containing street address
   */
  addressLine2: string | null;
  /**
   * City of the address.
   */
  city: string;
  /**
   * Latitude
   */
  latitude: number | null;
  /**
   * Longitude
   */
  longitude: number | null;
  /**
   * State for the address.
   */
  state: string;
  /**
   * Zip of the address.
   */
  zip: string;
}